Restructure benchmarks to use Criterion
- convert to a library
- make structs and functions public
- move benchmark code to separate file
- move setup functions to benchmark code
